src/python/requirements.txt
openai
PyGithub
pydantic
